blog list build issue fixed
