+++ src/app/blog/data.ts
@@ -1,0 +1,42 @@
+export const allBlogs = [
+  {
+    slug: "coderabbit-vs-others-ai-code-review-tools",
+    title: "CodeRabbit vs Others: Which is the Right AI Code Review Tool",
+    description: "CodeRabbit vs Other AI code review tools",
+    publishedAt: "2024-12-26T00:00:00Z",
+  },
+  {
+    slug: "stream-vs-twilio",
+    title: "Twilio vs Stream - A Guide to Building Smarter Support Bots",
+    description:
+      "A comparison of Twilio and Stream for building automated customer support solutions",
+    publishedAt: "2024-12-11T00:00:00Z",
+  },
+  {
+    slug: "supabase-vs-clerk",
+    title: "Supabase vs Clerk - Choose the Right Auth for Your SaaS",
+    description:
+      "An in-depth comparison of Clerk and Supabase to help you choose the best auth for your SaaS",
+    publishedAt: "2024-10-28T00:00:00Z",
+  },
+  {
+    slug: "mongoDB-vs-postgreSQL",
+    title: "MongoDB vs PostgreSQL- A Technical Comparison",
+    description:
+      "A technical comparison of MongoDB vs. PostgreSQL to help you choose the best database solution for your SaaS",
+    publishedAt: "2024-09-26T00:00:00Z",
+  },
+  {
+    slug: "state-of-databases-2024",
+    title: "State of Databases for Serverless in 2024",
+    description: "My thoughts on the state of Databases for Serverless",
+    publishedAt: "2024-09-03T00:00:00Z",
+  },
+  {
+    slug: "neon-vs-supabase",
+    title: "Neon Postgres vs Supabase: Choose the Right Database for Your SaaS",
+    description:
+      "An in-depth comparison of Neon and Supabase to help you choose the best database for your SaaS",
+    publishedAt: "2024-08-08T00:00:00Z",
+  },
+]
